# apps/api/openapi.internal.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "UsersController_deleteMe",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/users/me/contact-change/request": {
      "post": {
        "operationId": "UsersController_requestContactChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestContactChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/contact-change/verify": {
      "post": {
        "operationId": "UsersController_verifyContactChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyContactChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/contact-phone/request": {
      "post": {
        "operationId": "UsersController_requestContactPhoneChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestContactPhoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/contact-phone/verify": {
      "post": {
        "operationId": "UsersController_verifyContactPhoneChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyContactPhoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/profile": {
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/legal-consent": {
      "post": {
        "operationId": "UsersController_acceptLegalConsent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptLegalConsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/listings/{listingId}/media": {
      "get": {
        "operationId": "ListingMediaController_list",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "ListingMediaController_upload",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "AdminUsersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Avino API",
    "description": "API портала недвижимости Avino. Аутентификация: POST /auth/otp/request → /auth/otp/verify → { accessToken, refreshToken }; обновление — /auth/refresh; вход через Google — /auth/google; вход через Apple — /auth/apple. Bearer-токен передаётся в заголовке Authorization.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "default_language": {
            "type": "object"
          }
        }
      },
      "RequestContactChangeDto": {
        "type": "object",
        "properties": {
          "channel": {
            "type": "object"
          },
          "destination": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "channel",
          "destination"
        ]
      },
      "VerifyContactChangeDto": {
        "type": "object",
        "properties": {
          "channel": {
            "type": "object"
          },
          "destination": {
            "type": "string",
            "maxLength": 255
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "channel",
          "destination",
          "code"
        ]
      },
      "RequestContactPhoneDto": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "maxLength": 32
          }
        },
        "required": [
          "destination"
        ]
      },
      "VerifyContactPhoneDto": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "maxLength": 32
          },
          "code": {
            "type": "string",
            "minLength": 4,
            "maxLength": 8
          }
        },
        "required": [
          "destination",
          "code"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 100
          },
          "last_name": {
            "type": "string",
            "maxLength": 100
          },
          "display_name": {
            "type": "string",
            "maxLength": 150
          },
          "avatar_url": {
            "type": "string",
            "maxLength": 2048
          },
          "preferred_language": {
            "type": "object"
          }
        }
      },
      "AcceptLegalConsentDto": {
        "type": "object",
        "properties": {
          "terms_accepted": {
            "type": "boolean"
          },
          "privacy_accepted": {
            "type": "boolean"
          }
        },
        "required": [
          "terms_accepted",
          "privacy_accepted"
        ]
      },
...
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "object"
          },
          "reason": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}